requirements.txt
django==1.8.4
python-linkedin==4.2
python-social-auth
wikipedia==1.4.0
